+++ example/test/_unit_testing.dart
@@ -5,9 +5,6 @@
 import 'package:flutter_test/flutter_test.dart';
 
 import 'test_controller.dart';
-
-import 'test_listener.dart';
-
 import 'test_statex.dart';
 
 Future<void> unitTesting(WidgetTester tester) async {