+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 
-## 2.5.0+2
+## 2.5.0+3
 February 27, 2023
 - RecordExceptionMixin stores the last error
 - onAsyncError() no longer returns boolean
@@ -8,11 +8,13 @@ February 27, 2023
 - bool forEachState(void Function(StateX state) func, {bool? reversed}) {
 - startState and endState introduced to the StateX class
 - WidgetsBinding.instance in activate() & deactivate()
-- Added a State Listener and more tests for the forEach() & forEachState()
+- Incrementally adding tests for a higher percentage in Codecov
+- Updated example app in README.md
+- StateX.notifyClients() was corrected.
 
 ## 2.4.0
 February 21, 2023
-- Replaced buildWidget() function with buildF() function
+- Replace buildWidget() function with buildF() function
 
 ## 2.3.0
 February 17, 2023